Implement support for asyn requests using Celery
Async mode is now the one used by default.
Maximum concurrency of celery workers can be set in the docker-compose file.

https://github.com/AudioCommons/ac-mediator/issues/5

+++ api/response_aggregator.py
@@ -121,7 +121,8 @@ class ResponseAggregator(object):
         if response is None:
             return None
         to_return = response.copy()
-        if response['status'] == RESPONSE_STATUS_FINISHED:
+        to_return.update({'response_id': response_id})
+        if response['status'] == RESPONSE_STATUS_FINISHED and settings.DELETE_RESPONSES_AFTER_CONSUMED:
             self.store.delete_response(response_id)  # If response has been all loaded, delete it from pool
         return to_return
